test filtering + worker shared dir + file check + preprocessing lock

+++ lemontest/test_scheduler/test_scheduler.py
@@ -1,13 +1,21 @@
+import shutil
 from classes.test_scheduler import AbstractScheduler
+from classes.test import AbstractTest
 from test_worker.test_worker import TestWorker
+from util.util import die
+
+# typing
+from argparse import Namespace
+from typing import Dict, Any
 
 # apply pool patch
 import util.istarmap
-from multiprocessing import Pool, log_to_stderr, SUBDEBUG 
+from multiprocessing import set_start_method, Pool, Lock, log_to_stderr, SUBDEBUG
+from termcolor import colored as termcolor_colored
 
-import sys
 import tqdm
-import subprocess
+import glob
+import tempfile
 
 """
 import logging
@@ -16,33 +24,63 @@ logger.setLevel(SUBDEBUG)
 """
 
 class TestScheduler(AbstractScheduler):
+    args = None
     parameters = None
+    shared_dir = None
     worker_pool = None
 
-    def __init__(self, **parameters):
+    def __init__(self, args: Namespace, parameters: Dict[str, Any]):
+        self.args = args
         self.parameters = parameters
+        self.shared_dir = tempfile.mkdtemp()
+        self.colored = (
+            termcolor_colored
+            if parameters["colorize_output"]
+            else lambda x, *a, **kw: x
+        )
         try:
+            # set the fork start method
+            # TODO: kill windows support at launch (lemontest.py)
+            set_start_method('fork')
+            # spawn Lock for test preprocessing shared directory access
+            pLock = Lock()
             # spawn worker pool
-            self.worker_pool = Pool(self.parameters["worker_count"], maxtasksperchild=1)
+            self.worker_pool = Pool(initializer=test_worker_init, initargs=(pLock,), processes=self.parameters["worker_count"], maxtasksperchild=1)
         except Exception as err:
             # cleanup worker pool
             if self.worker_pool:
                 self.worker_pool.terminate() # send SIGTERM to worker processes
-            print(err)
-            sys.exit(1)
+            die(err)
 
     def __str__(self):
         pass
 
-    def schedule(self, tests):
-        tests = [(test, self.parameters) for test in tests]
+    def schedule(self, tests: Dict[str, AbstractTest]):
+        # check there are tests available to run
+        if not tests:
+            die(f"autotest not available for {self.args.exercise}")
+        if not self.args.labels:
+            die("nothing to test")
+
+        # process tests to be run
+        tests = [(test, self.shared_dir, self.parameters) for test in tests]
+
+        # FIXME: compile.sh and runtests.pl was executed here usually
+
+        # check tests to ensure we have all files to execute tests, otherwise die with missing files
+        req_files_set = set.intersection(
+            *[set(test.params()["files"]) for (test, _, _) in tests]
+        )
+        missing_files = [f for f in req_files_set if not glob.glob(f)]
+        if missing_files:
+            die(f"Unable to run tests because these files were missing: {self.colored(' '.join(missing_files), 'red')}")
 
         # schedule tests for execution and show progress
         test_res = list(tqdm.tqdm(self.worker_pool.istarmap(test_worker, tests, chunksize=1), total=len(tests), desc=f"Running {len(tests)} tests:", unit=" tests"))
         
         # terminate worker pool as work is done
         self.worker_pool.terminate()
-        self.worker_pool.join() # this doesn't sometimes work with .close so just terminate
+        self.worker_pool.join() # this doesn't sometimes work with .close() so just terminate
 
         # reset worker_pool in case cleanup gets triggered by signal
         self.worker_pool = None
@@ -53,13 +91,23 @@ class TestScheduler(AbstractScheduler):
     def cleanup(self):
         if self.worker_pool:
             self.worker_pool.terminate() # send SIGTERM to worker processes
+        if self.shared_dir:
+            shutil.rmtree(self.shared_dir)
+
+# setup a global variable to inherit a global lock for the test preprocessing
+# see: test_worker to see why this is absolutely necessary
+# we can't use multiprocessing.Manager because that will break if we
+# isolate networking within the sandbox (more common than not)
+def test_worker_init(lock):
+    global pLock
+    pLock = lock
 
 
 # initalise worker and execute task
-def test_worker(test, parameters):
-    worker = TestWorker(**parameters)
+def test_worker(test, shared_dir, parameters):
+    worker = TestWorker(shared_dir, **parameters)
     worker.setup()
-    res = worker.execute(test)
+    res = worker.execute(test, pLock) #pLock available from Pool initializer (global var)
     worker.cleanup()
     return res
